refactor: move json schema generators to a separate module
